fix organization destroy callback on business_profile

+++ app/models/business_profile.rb
@@ -23,7 +23,7 @@ class BusinessProfile < ActiveRecord::Base
   validates_associated :content
 
   after_destroy do
-    organization.destroy
+    organization.destroy if organization.present?
   end
 
   belongs_to :business_location, dependent: :destroy
